Extracted base `HoodieTableInputFormat`
